refactor: extract helper functions into separate file
